Fix bugs in question models

+++ app/api/models/questions.py
@@ -85,10 +85,17 @@ class Question(User, DatabaseConnection):
         except Exception as e:
             logging.error(e)
             return make_response(jsonify({'message': str(e)}), 500)
+    
     @staticmethod
-    def delete_question(user_id, qtn_id):
+    def delete_question(qtn_id, user_id):
         with DatabaseConnection() as cursor:
                 try:
+                    query = "SELECT FROM questions WHERE qtn_id = '%s'" % qtn_id
+                    cursor.execute(query)
+                    question = cursor.fetchone()
+                    print(question)
+                    if not question:
+                        return {"message": "Question doesn't exist"}
                     sql = "DELETE FROM questions WHERE qtn_id = %s AND user_id = %s"
                     cursor.execute(sql, [qtn_id, user_id])
                     return {"message": "Question deleted"}
